+++ ZalithLauncher/src/main/java/com/movtery/zalithlauncher/feature/mod/modloader/OptiFineDownloadTask.kt
@@ -35,7 +35,7 @@ class OptiFineDownloadTask(
         return mDestinationFile
     }
 
-    override fun updateProgress(curr: Int, max: Int) {
+    override fun updateProgress(curr: Long, max: Long) {
         val progress100 = ((curr.toFloat() / max.toFloat()) * 100f).toInt()
         ProgressKeeper.submitProgress(
             ProgressLayout.INSTALL_RESOURCE,
